% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ageClass.R
\name{ageClass}
\alias{ageClass}
\title{ageClass}
\usage{
ageClass(ageVector, classType = "class10")
}
\arguments{
\item{ageVector}{numeric vector including the age to be proccess}

\item{classType}{charactor vector of length 1 used to specify which age class cut to apply}
}
\description{
Function to create variables with different age class
}
\examples{
y <- data.frame(id=1:10, age=sample(18:60, 10))
y <- cbind(y, ageClass(ageVector = y$age, classType = "class10"))  # classType \%in\% c("classAMT", "class10", "class5", "classDecennie")
}

