% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetimeValorisation2.R
\name{datetimeValorisation2}
\alias{datetimeValorisation2}
\title{datetimeValorisation2}
\usage{
datetimeValorisation2(input, variableName, workOrder,
  simplifiedVector = FALSE)
}
\arguments{
\item{input}{a data frame or a vector : data frame must be in this format : id, datetimesee. See simplifiedVector for info on the format of the vector option}

\item{variableName}{single length vector to indicate the prefix of the newly create variable in this format : paste0(variableName, "dependingOnWorkOrder")}

\item{workOrder}{vector to include all transformation to do on the datetime object : c("year", "month", "day", "week", "wday", "ymd", "ym", "yw")}

\item{simplifiedVector}{if simplifiedVector is TRUE, then the data frame is in a vector format containing the datetimes}
}
\description{
Function to enrich a datetime object
}
\examples{
x <- data.frame(x1 = 1:3, x2 = as.Date(c("2013-06-20", "2013-08-15", "2015-01-02")))
y <- datetimeValorisation2(x, "TypeA", c("year", "month", "day", "week", "wday", "ymd", "ym", "yw"), simplifiedVector = FALSE)
}

