% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfFromListToMainEnv.R
\name{dfFromListToMainEnv}
\alias{dfFromListToMainEnv}
\title{dfFromListToMainEnv}
\usage{
dfFromListToMainEnv(dfList, dfNames = NULL)
}
\arguments{
\item{dfList}{A list object containing data frames}

\item{dfNames}{Optionnal argument to be specified if dfList doesn't have proper names assigned to its data frames}
}
\description{
Function to convert data frames contained in a list object into data frame objects in the main environment
}
\examples{
dfFromListToMainEnv(dfList = list(data.frame(x1=1:10, x2=11:20), data.frame(x3=1:3)), dfNames = c("foo", "bar"))
}

