% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthdist.R
\name{earth.dist}
\alias{earth.dist}
\title{earth.dist}
\usage{
earth.dist(long1, lat1, long2, lat2)
}
\arguments{
\item{long1}{longitude of the first coordinate}

\item{lat1}{latitude of the first coordinate}

\item{long2}{longitude of the second coordinate}

\item{lat2}{latitude of the second coordinate}
}
\description{
Function to calculate the metric distance between two coordinates (lat, lon)
}
\examples{
earth.dist(-70.34, 43.23, -71.29, 43.21)
}

