% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAllCsvFromFolder.R
\name{importAllCsvFromFolder}
\alias{importAllCsvFromFolder}
\title{importAllCsvFromFolder}
\usage{
importAllCsvFromFolder(folder = getwd())
}
\arguments{
\item{folder}{folder containing csv files by default it is set to the working directory}
}
\description{
Function to import all csv files contained inside a specified folder
}
\examples{
importAllCsvFromFolder(folder = "Data")
}

