% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliderValues.R
\name{sliderValues}
\alias{sliderValues}
\title{sliderValues}
\usage{
sliderValues(inputId, label, values, from, to = NULL, grid = TRUE,
  width = NULL, postfix = NULL, prefix = NULL, dragRange = TRUE,
  disable = FALSE, animate = FALSE)
}
\description{
Function to make a custom slider to navigate through custom labels
}
\examples{
importPostgreSQLTable("db", "localhost", 5431, "postgres", "pass", c("table1", "table2"), c(0.5, 1), c(TRUE, FALSE))
sliderValues(inputId = "test", label = "Month", width = "100\%",values = choices_month[4:6], from = choices_month[4], to = choices_month[6],grid = FALSE, animate = animationOptions(interval = 1500))
}

